DATAJPA-444 - Prefer Hibernate 4.3 provider interface over legacy one.
Accessing the legacy persistence provider class in Hibernate 4.3 causes a warning being logged. As we generally want to detect the presence of Hibernate in general only, we now check for the Hibernate 4.3 persistence provider interface first to immediately find the new interface and thus avoid the warning.

Upgraded to Hibernate 4.3.1.