DATAJPA-483 - Fixed parameter binding detection with parentheses.
When a parameter was listed with parentheses we didn't detect a custom binding and fell back to the standard binding. This effectively disabled the array-to-collection binding which is currently necessary for in bindings as some persistence providers do not bind arrays to in-clauses correctly.

Tweaked the regular expression to detect the bindings to accept the optional parentheses.