DATAJPA-504 - Regression in Querydsl OrderSpecifier customization.
Removed custom querydsl left join generation for order fields.
The recent versions of Querydsl already do this properly - so there is no need for us anymore to add left-joins ourselves.
Previously we had to generate additional left-joins since Querydsl did not support ordering by an attribute that was not part of the projection.

Original pull request: #73.