DATAJPA-929 - Further guards against null values returned from ManagedType.
Extracted the functionality to check for JPA managed types into a dedicated wrapper for the JPA Metamodel. This allows all clients to benefit from the null guards we have to put in place due to [0]. Adapted the fix introduced for DATAJPA-904 to make use of the newly created infrastructure, too.

[0] https://hibernate.atlassian.net/browse/HHH-10968