DATAJPA-998 - Prepare 2.0 development.
Upgraded to Spring Data build parent and Commons in 2.0. No upgrade to Spring 5 yet as it requires Hibernate 5 which currently breaks our tests, mostly because of [0].

Added workaround for compiler ambiguities that stem from the upgrade to 1.8 as compile target [1]. Temporarily removed tests for JpaSort for the same reason.

Adapted to API changes in DefaultResourceLoader in preparation of the move to Spring 5.

[0] https://hibernate.atlassian.net/browse/HHH-10690
[1] http://mail.openjdk.java.net/pipermail/compiler-dev/2016-November/010521.html