DATAJPA-1140 - Allow mixing of SpEL expression and normal query parameters.
Restructured the complete parameter binding. There are now two principle ways to perform the binding:

- if no query is known to the binder the method parameters get iterated and assigned to query parameters.
This is implemented by the bind method of the ParameterBinder.
- if a query is known all the bindings, i.e the query parameters are iterated and matching values are searched.
This is implemented in the QueryAwareParameterBinder.

For finding a suitable value there exists various QueryParameterSetterStrategies. The code calling the various Query.setParameter(…) overloads is extracted and gathered in QueryParameterSetter.

Original pull request: #206.