DATAJPA-1248 - Fixed build for Hibernate 5.3.
Removed obsolete version detection and associated persistence units from tests.

Added explicite cascade option to many-to-one reference in id. It is not clear if this is a bug in Hibernate or not. See HHH-12251 for details.

The fix of HHH-12119 made it obvious that our use of Tuples for Projections from native queries were not correct. When looking up Tuple values we now find them using a property name with correct upper/lower case as well as with the upper case as it is actually returned by the JDBC driver.

Converted all touched tests using Hamcrest to use AssertJ instead.

See also: 
https://hibernate.atlassian.net/browse/HHH-12251
https://hibernate.atlassian.net/browse/HHH-12119

Original pull request: #245.