DATAJPA-1320 - Adapt to API changes in Spring Data Commons.
JpaPersistentProperty now exposes an ….isEmbeddable(), currently used in JpaMetamodelMappingContext implementation to make sure that embeddable types are traversed when it comes to PersistentPropertyPath detection. That in turn is now used for the detection of auditing metdata, so that auditing annotations can also be used on embeddables.

Lazified the decision whether a property is an association which prevents repeated annotation lookups.

Related tickets: DATACMNS-1275.