DATAJPA-1394 - EntityPathResolver selection is now done via ObjectProvider.
Instead of manually triggering a beans-of-type lookup via BeanLookup, we now rather get an ObjectProvider for EntityPathResolver injected and rely on its ….getIfAvailable(…) semantics to fallback to the default or reject the lookup if multiple instances were registered.

Changed the test cases to integration ones as we now need to verify the container interaction.

Related ticket: DATAJPA-1234.