DATAJPA-1448 - Properly avoid double registration of AnnotationBeanConfigurerAspect.
Previously we used the aspect's class name to guard the registration of a AnnotationBeanConfigurerAspect bean but eventually ended up registering it under the globally used name (org.springframework.context.config.internalBeanConfigurerAspect).

That caused our bean being registered even if there already was one in the context causing the original bean to be overridden. The latter is now disabled by default in Spring Boot 2.1 so that it causes applications failing to start.

We now properly check for the correct bean name to guard the registration.