DATAJPA-872 - Allow for duplicate hints.
Some JPA implementations allow to have multiple hints for the same key/name.
The former implementation did not support that since it stored hints as Maps.

With this change they are now stored as lists and duplicates are passed on to the JPA implementation.

How the implementation deals with such a situation depends on that implementation.

Original pull request: #426.